Move slack-ping state to slack-ping componenet
